breaking: move remote function types and isValidationError to @sveltejs/kit/remote
