+++ .changeset/bright-trees-listen.md
@@ -1,0 +1,5 @@
+---
+"@sveltejs/kit": major
+---
+
+breaking: move remote function types and `isValidationError` to `@sveltejs/kit/remote`
