+++ documentation/docs/98-reference/15-@sveltejs-kit-remote.md
@@ -1,0 +1,5 @@
+---
+title: @sveltejs/kit/remote
+---
+
+> MODULE: @sveltejs/kit/remote
