+++ packages/kit/scripts/generate-dts.js
@@ -9,6 +9,7 @@ await createBundle({
 		'@sveltejs/kit/hooks': 'src/exports/hooks/public.d.ts',
 		'@sveltejs/kit/node': 'src/exports/node/index.js',
 		'@sveltejs/kit/params': 'src/exports/params/public.d.ts',
+		'@sveltejs/kit/remote': 'src/exports/remote/public.d.ts',
 		'@sveltejs/kit/vite': 'src/exports/vite/index.js',
 		'$app/env': 'src/runtime/app/env/types.d.ts',
 		'$app/forms': 'src/runtime/app/forms/public.d.ts',
