+++ packages/kit/src/runtime/app/server/remote/form.js
@@ -1,4 +1,4 @@
-/** @import { RemoteFormInput, RemoteForm, InvalidField } from '$app/server' */
+/** @import { RemoteFormInput, RemoteForm, InvalidField } from '@sveltejs/kit/remote' */
 /** @import { InternalRemoteFormIssue, MaybePromise, HasNonOptionalBoolean, RemoteFormInternals } from 'types' */
 /** @import { StandardSchemaV1 } from '@standard-schema/spec' */
 import { get_request_store } from '@sveltejs/kit/internal/server';
