+++ packages/kit/src/runtime/client/remote-functions/prerender.svelte.js
@@ -1,4 +1,4 @@
-/** @import { RemotePrerenderFunction } from '$app/server' */
+/** @import { RemotePrerenderFunction } from '@sveltejs/kit/remote' */
 import { app_dir, base } from '#app/paths';
 import { version } from '$app/env';
 import * as devalue from 'devalue';
