+++ packages/kit/src/runtime/client/remote-functions/query-batch.svelte.js
@@ -1,4 +1,4 @@
-/** @import { RemoteQueryFunction } from '$app/server' */
+/** @import { RemoteQueryFunction } from '@sveltejs/kit/remote' */
 import { app_dir, base } from '#app/paths';
 import { _goto } from '../client.js';
 import { QUERY_FUNCTION_ID, remote_request } from './shared.svelte.js';
