+++ packages/kit/test/apps/async/src/routes/remote/query-runtime-errors/not-tracked/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { RemoteQuery } from '$app/server';
+	import type { RemoteQuery } from '@sveltejs/kit/remote';
 	import { get_count } from '../../query-command.remote.js';
 
 	let status = $state('idle');
