+++ packages/kit/tsconfig.json
@@ -11,6 +11,7 @@
 			"@sveltejs/kit/env": ["./src/exports/env/public.d.ts"],
 			"@sveltejs/kit/hooks": ["./src/exports/hooks/public.d.ts"],
 			"@sveltejs/kit/params": ["./src/exports/params/public.d.ts"],
+			"@sveltejs/kit/remote": ["./src/exports/remote/public.d.ts"],
 			"@sveltejs/kit/vite": ["./src/exports/vite/index.js"],
 			"@sveltejs/kit/node": ["./src/exports/node/index.js"],
 			"@sveltejs/kit/internal": ["./src/exports/internal/index.js"],
